Move model mapping helpers to standalone class.
