+++ src/Libraries/Microsoft.Extensions.AI.OpenAI/OpenAIJsonContext.cs
@@ -1,6 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Collections.Generic;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
@@ -11,6 +12,7 @@ namespace Microsoft.Extensions.AI;
     UseStringEnumConverter = true,
     DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
     WriteIndented = true)]
-[JsonSerializable(typeof(OpenAIChatClient.OpenAIChatToolJson))]
 [JsonSerializable(typeof(OpenAIRealtimeExtensions.ConversationFunctionToolParametersSchema))]
+[JsonSerializable(typeof(OpenAIModelMappers.OpenAIChatToolJson))]
+[JsonSerializable(typeof(IDictionary<string, object?>))]
 internal sealed partial class OpenAIJsonContext : JsonSerializerContext;
